src/msb-mqtt/requirements.txt
paho-mqtt==1.5.1
pytz==2020.5
